feat(web): add landing page and move app route to /app
Serve a public homepage at / with concise product positioning and login CTA, and relocate the existing authenticated app experience to /app to support marketing-site entry on the new custom domain. Closes #94.

Made-with: Cursor
